Moved around package layout
